requirements.txt
pygame
neat-python
numpy
graphviz
matplotlib 